feat: Integrate MPC into RL training pipeline, using RL planned output `[vx, vy, yawrate, height]`.
* File `legged_robot.py`:
(1) The initialization of some arrays are changed from using `self.num_actions` to `self.num_dofs` since the number of actions may not be the number of the robot's leg dof.
(2) The `step` function is also modified to receive both the actions from RL and the torques from MPC. RL actions is updated and stored into self.actions here.
Both these two changes are not expected to affect other existing tasks.